src/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
SQLAlchemy>=2.0.28
Flask-JWT-Extended==4.4.4
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
pybreaker==1.1.0
marshmallow==3.20.2
structlog==24.1.0
Werkzeug==3.0.1
prometheus-client==0.20.0
bootstrap-flask==2.3.0
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
pytest==8.0.0
Flask-WTF==1.2.1
email-validator==2.1.1
